requirements.txt
streamlit
langchain
langchain-core
langchain-google-genai
langchain-tavily
langchain-google-genai
geopy
requests
uuid
python-dotenv